+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
  - Fix settings menu layout
  - Significantly improve speed of dependency-checking
+ - Fix missing names for Function Actions
 
 ### Features
 
  - Add dependency system for Function Actions
  - Add proper dependencies for infrared Send Function Action
+ - Improve Infrared Sed Action by detecting remotes and codes
 
 
 ## 7.3.0 (2019-02-22)